requirements.txt
streamlit
altair
tornado
click
watchdog
pypdf
python-dotenv
openai-agents
